vsiaz.py: separate tests on real instance in a separate file
