+++ autotest/gcore/vsiaz.py
@@ -28,7 +28,6 @@
 # DEALINGS IN THE SOFTWARE.
 ###############################################################################
 
-import stat
 import sys
 from osgeo import gdal
 
@@ -37,6 +36,7 @@ import gdaltest
 import webserver
 import pytest
 
+pytestmark = pytest.mark.skipif(not gdaltest.built_against_curl(), reason="GDAL not built against curl")
 
 def open_for_read(uri):
     """
@@ -45,126 +45,17 @@ def open_for_read(uri):
     return gdal.VSIFOpenExL(uri, 'rb', 1)
 
 ###############################################################################
+@pytest.fixture(autouse=True, scope='module')
+def startup_and_cleanup():
 
-
-def test_vsiaz_init():
-
-    gdaltest.az_vars = {}
+    az_vars = {}
     for var in ('AZURE_STORAGE_CONNECTION_STRING', 'AZURE_STORAGE_ACCOUNT',
                 'AZURE_STORAGE_ACCESS_KEY', 'AZURE_SAS', 'AZURE_NO_SIGN_REQUEST'):
-        gdaltest.az_vars[var] = gdal.GetConfigOption(var)
-        if gdaltest.az_vars[var] is not None:
+        az_vars[var] = gdal.GetConfigOption(var)
+        if az_vars[var] is not None:
             gdal.SetConfigOption(var, "")
 
     assert gdal.GetSignedURL('/vsiaz/foo/bar') is None
-
-###############################################################################
-# Error cases
-
-
-def test_vsiaz_real_server_errors():
-
-    if not gdaltest.built_against_curl():
-        pytest.skip()
-
-    # Missing AZURE_STORAGE_ACCOUNT
-    gdal.ErrorReset()
-    with gdaltest.error_handler():
-        f = open_for_read('/vsiaz/foo/bar')
-    assert f is None and gdal.VSIGetLastErrorMsg().find('AZURE_STORAGE_ACCOUNT') >= 0
-
-    gdal.ErrorReset()
-    with gdaltest.error_handler():
-        f = open_for_read('/vsiaz_streaming/foo/bar')
-    assert f is None and gdal.VSIGetLastErrorMsg().find('AZURE_STORAGE_ACCOUNT') >= 0
-
-    # Invalid AZURE_STORAGE_CONNECTION_STRING
-    with gdaltest.config_option('AZURE_STORAGE_CONNECTION_STRING', 'invalid'):
-        gdal.ErrorReset()
-        with gdaltest.error_handler():
-            f = open_for_read('/vsiaz/foo/bar')
-        assert f is None
-
-    # Missing AZURE_STORAGE_ACCESS_KEY
-    gdal.ErrorReset()
-    with gdaltest.config_options({'AZURE_STORAGE_ACCOUNT': 'AZURE_STORAGE_ACCOUNT',
-                                  'CPL_AZURE_VM_API_ROOT_URL': 'disabled'}):
-        with gdaltest.error_handler():
-            f = open_for_read('/vsiaz/foo/bar')
-        assert f is None and gdal.VSIGetLastErrorMsg().find('AZURE_STORAGE_ACCESS_KEY') >= 0
-
-    # AZURE_STORAGE_ACCOUNT and AZURE_STORAGE_ACCESS_KEY but invalid
-    gdal.ErrorReset()
-    with gdaltest.config_options({'AZURE_STORAGE_ACCOUNT': 'AZURE_STORAGE_ACCOUNT',
-                                  'AZURE_STORAGE_ACCESS_KEY': 'AZURE_STORAGE_ACCESS_KEY'}):
-        with gdaltest.error_handler():
-            f = open_for_read('/vsiaz/foo/bar.baz')
-        if f is not None:
-            if f is not None:
-                gdal.VSIFCloseL(f)
-            if gdal.GetConfigOption('APPVEYOR') is not None:
-                return
-            pytest.fail(gdal.VSIGetLastErrorMsg())
-
-        gdal.ErrorReset()
-        with gdaltest.error_handler():
-            f = open_for_read('/vsiaz_streaming/foo/bar.baz')
-        assert f is None, gdal.VSIGetLastErrorMsg()
-
-
-###############################################################################
-# Test AZURE_NO_SIGN_REQUEST=YES
-
-
-def test_vsiaz_no_sign_request():
-
-    if not gdaltest.built_against_curl():
-        pytest.skip()
-
-    with gdaltest.config_options({ 'AZURE_STORAGE_ACCOUNT': 'naipblobs', 'AZURE_NO_SIGN_REQUEST': 'YES'}):
-        actual_url = gdal.GetActualURL('/vsiaz/naip/v002/al/2015/al_100cm_2015/30086/m_3008601_ne_16_1_20150804.tif')
-        assert actual_url == 'https://naipblobs.blob.core.windows.net/naip/v002/al/2015/al_100cm_2015/30086/m_3008601_ne_16_1_20150804.tif'
-        assert actual_url == gdal.GetSignedURL('/vsiaz/naip/v002/al/2015/al_100cm_2015/30086/m_3008601_ne_16_1_20150804.tif')
-
-        f = open_for_read('/vsiaz/naip/v002/al/2015/al_100cm_2015/30086/m_3008601_ne_16_1_20150804.tif')
-        if f is None:
-            if gdaltest.gdalurlopen('https://naipblobs.blob.core.windows.net/naip/v002/al/2015/al_100cm_2015/30086/m_3008601_ne_16_1_20150804.tif') is None:
-                pytest.skip('cannot open URL')
-            pytest.fail()
-
-        gdal.VSIFCloseL(f)
-
-        assert 'm_3008601_ne_16_1_20150804.tif' in gdal.ReadDir('/vsiaz/naip/v002/al/2015/al_100cm_2015/30086/')
-
-###############################################################################
-# Test AZURE_SAS option
-
-
-def test_vsiaz_sas():
-
-    if not gdaltest.built_against_curl():
-        pytest.skip()
-
-    # See https://azure.microsoft.com/en-us/services/open-datasets/catalog/naip/ for the value of AZURE_SAS
-    with gdaltest.config_options({ 'AZURE_STORAGE_ACCOUNT': 'naipblobs', 'AZURE_SAS': 'st=2019-07-18T03%3A53%3A22Z&se=2035-07-19T03%3A53%3A00Z&sp=rl&sv=2018-03-28&sr=c&sig=2RIXmLbLbiagYnUd49rgx2kOXKyILrJOgafmkODhRAQ%3D'}):
-        actual_url = gdal.GetActualURL('/vsiaz/naip/v002/al/2015/al_100cm_2015/30086/m_3008601_ne_16_1_20150804.tif')
-        assert actual_url == 'https://naipblobs.blob.core.windows.net/naip/v002/al/2015/al_100cm_2015/30086/m_3008601_ne_16_1_20150804.tif'
-        assert gdal.GetSignedURL('/vsiaz/naip/v002/al/2015/al_100cm_2015/30086/m_3008601_ne_16_1_20150804.tif') == 'https://naipblobs.blob.core.windows.net/naip/v002/al/2015/al_100cm_2015/30086/m_3008601_ne_16_1_20150804.tif?st=2019-07-18T03%3A53%3A22Z&se=2035-07-19T03%3A53%3A00Z&sp=rl&sv=2018-03-28&sr=c&sig=2RIXmLbLbiagYnUd49rgx2kOXKyILrJOgafmkODhRAQ%3D'
-
-        f = open_for_read('/vsiaz/naip/v002/al/2015/al_100cm_2015/30086/m_3008601_ne_16_1_20150804.tif')
-        if f is None:
-            if gdaltest.gdalurlopen('https://naipblobs.blob.core.windows.net/naip/v002/al/2015/al_100cm_2015/30086/m_3008601_ne_16_1_20150804.tif') is None:
-                pytest.skip('cannot open URL')
-            pytest.fail()
-
-        gdal.VSIFCloseL(f)
-
-        assert 'm_3008601_ne_16_1_20150804.tif' in gdal.ReadDir('/vsiaz/naip/v002/al/2015/al_100cm_2015/30086/')
-
-###############################################################################
-
-
-def test_vsiaz_start_webserver():
 
     gdaltest.webserver_process = None
     gdaltest.webserver_port = 0
@@ -181,6 +72,17 @@ def test_vsiaz_start_webserver():
     gdal.SetConfigOption('AZURE_STORAGE_ACCOUNT', '')
     gdal.SetConfigOption('AZURE_STORAGE_ACCESS_KEY', '')
     gdal.SetConfigOption('CPL_AZURE_TIMESTAMP', 'my_timestamp')
+
+    yield
+
+    # Clearcache needed to close all connections, since the Python server
+    # can only handle one connection at a time
+    gdal.VSICurlClearCache()
+
+    webserver.server_stop(gdaltest.webserver_process, gdaltest.webserver_port)
+
+    for var in az_vars:
+        gdal.SetConfigOption(var, az_vars[var])
 
 ###############################################################################
 # Test with a fake Azure Blob server
@@ -1348,149 +1250,3 @@ def test_vsiaz_read_credentials_simulated_azure_vm_expiration():
             gdal.VSIFCloseL(f)
 
         assert data == 'foo'
-
-
-###############################################################################
-
-
-def test_vsiaz_stop_webserver():
-
-    if gdaltest.webserver_port == 0:
-        pytest.skip()
-
-    # Clearcache needed to close all connections, since the Python server
-    # can only handle one connection at a time
-    gdal.VSICurlClearCache()
-
-    webserver.server_stop(gdaltest.webserver_process, gdaltest.webserver_port)
-
-###############################################################################
-# Nominal cases (require valid credentials)
-
-
-def test_vsiaz_extra_1():
-
-    if not gdaltest.built_against_curl():
-        pytest.skip()
-
-    az_resource = gdal.GetConfigOption('AZ_RESOURCE')
-    if az_resource is None:
-        pytest.skip('Missing AZ_RESOURCE')
-
-    if '/' not in az_resource:
-        path = '/vsiaz/' + az_resource
-        statres = gdal.VSIStatL(path)
-        assert statres is not None and stat.S_ISDIR(statres.mode), \
-            ('%s is not a valid bucket' % path)
-
-        readdir = gdal.ReadDir(path)
-        assert readdir is not None, 'ReadDir() should not return empty list'
-        for filename in readdir:
-            if filename != '.':
-                subpath = path + '/' + filename
-                assert gdal.VSIStatL(subpath) is not None, \
-                    ('Stat(%s) should not return an error' % subpath)
-
-        unique_id = 'vsiaz_test'
-        subpath = path + '/' + unique_id
-        ret = gdal.Mkdir(subpath, 0)
-        assert ret >= 0, ('Mkdir(%s) should not return an error' % subpath)
-
-        readdir = gdal.ReadDir(path)
-        assert unique_id in readdir, \
-            ('ReadDir(%s) should contain %s' % (path, unique_id))
-
-        ret = gdal.Mkdir(subpath, 0)
-        assert ret != 0, ('Mkdir(%s) repeated should return an error' % subpath)
-
-        ret = gdal.Rmdir(subpath)
-        assert ret >= 0, ('Rmdir(%s) should not return an error' % subpath)
-
-        readdir = gdal.ReadDir(path)
-        assert unique_id not in readdir, \
-            ('ReadDir(%s) should not contain %s' % (path, unique_id))
-
-        ret = gdal.Rmdir(subpath)
-        assert ret != 0, ('Rmdir(%s) repeated should return an error' % subpath)
-
-        ret = gdal.Mkdir(subpath, 0)
-        assert ret >= 0, ('Mkdir(%s) should not return an error' % subpath)
-
-        f = gdal.VSIFOpenL(subpath + '/test.txt', 'wb')
-        assert f is not None
-        gdal.VSIFWriteL('hello', 1, 5, f)
-        gdal.VSIFCloseL(f)
-
-        ret = gdal.Rmdir(subpath)
-        assert ret != 0, \
-            ('Rmdir(%s) on non empty directory should return an error' % subpath)
-
-        f = gdal.VSIFOpenL(subpath + '/test.txt', 'rb')
-        assert f is not None
-        data = gdal.VSIFReadL(1, 5, f).decode('utf-8')
-        assert data == 'hello'
-        gdal.VSIFCloseL(f)
-
-        assert gdal.Rename(subpath + '/test.txt', subpath + '/test2.txt') == 0
-
-        f = gdal.VSIFOpenL(subpath + '/test2.txt', 'rb')
-        assert f is not None
-        data = gdal.VSIFReadL(1, 5, f).decode('utf-8')
-        assert data == 'hello'
-        gdal.VSIFCloseL(f)
-
-        ret = gdal.Unlink(subpath + '/test2.txt')
-        assert ret >= 0, \
-            ('Unlink(%s) should not return an error' % (subpath + '/test2.txt'))
-
-        ret = gdal.Rmdir(subpath)
-        assert ret >= 0, ('Rmdir(%s) should not return an error' % subpath)
-
-        return
-
-    f = open_for_read('/vsiaz/' + az_resource)
-    assert f is not None
-    ret = gdal.VSIFReadL(1, 1, f)
-    gdal.VSIFCloseL(f)
-
-    assert len(ret) == 1
-
-    # Same with /vsiaz_streaming/
-    f = open_for_read('/vsiaz_streaming/' + az_resource)
-    assert f is not None
-    ret = gdal.VSIFReadL(1, 1, f)
-    gdal.VSIFCloseL(f)
-
-    assert len(ret) == 1
-
-    if False:  # pylint: disable=using-constant-test
-        # we actually try to read at read() time and bSetError = false
-        # Invalid bucket : "The specified bucket does not exist"
-        gdal.ErrorReset()
-        f = open_for_read('/vsiaz/not_existing_bucket/foo')
-        with gdaltest.error_handler():
-            gdal.VSIFReadL(1, 1, f)
-        gdal.VSIFCloseL(f)
-        assert gdal.VSIGetLastErrorMsg() != ''
-
-    # Invalid resource
-    gdal.ErrorReset()
-    f = open_for_read('/vsiaz_streaming/' + az_resource + '/invalid_resource.baz')
-    assert f is None, gdal.VSIGetLastErrorMsg()
-
-    # Test GetSignedURL()
-    signed_url = gdal.GetSignedURL('/vsiaz/' + az_resource)
-    f = open_for_read('/vsicurl_streaming/' + signed_url)
-    assert f is not None
-    ret = gdal.VSIFReadL(1, 1, f)
-    gdal.VSIFCloseL(f)
-
-    assert len(ret) == 1
-
-###############################################################################
-
-
-def test_vsiaz_cleanup():
-
-    for var in gdaltest.az_vars:
-        gdal.SetConfigOption(var, gdaltest.az_vars[var])
